services/ml/production/cycle.py
"""
Module pour l'optimisation des cycles de culture.
"""

from typing import Dict, Any, List, Optional
from datetime import date, timedelta
import numpy as np
from sqlalchemy.orm import Session

from .base import BaseProductionML
from models.production import CultureType, Parcelle
from services.weather_service import WeatherService

class CycleOptimizer(BaseProductionML):
    """Service ML pour l'optimisation des cycles de culture"""
    
    def __init__(self, db: Session):
        super().__init__(db)
        self.weather_service = WeatherService(db)

    async def optimize_cycle_culture(
        self,
        parcelle_id: str,
        date_debut: Optional[date] = None
    ) -> Dict[str, Any]:
        """Optimise le planning d'un cycle de culture"""
        # Récupération des données
        parcelle = self.db.query(Parcelle).get(parcelle_id)
        historique = await self._get_historique_cycles(parcelle_id)
        previsions_meteo = await self.weather_service.get_forecast(parcelle_id)
        
        # Calcul de la date optimale de début si non spécifiée
        if not date_debut:
            date_debut = await self._calculate_optimal_start_date(
                parcelle.culture_type,
                previsions_meteo
            )
            
        # Optimisation des étapes du cycle
        etapes = await self._optimize_cycle_steps(
            parcelle.culture_type,
            date_debut,
            previsions_meteo
        )
        
        return {
            "date_debut_optimale": date_debut,
            "date_fin_prevue": etapes[-1]["date_fin"],
            "etapes": etapes,
            "rendement_prevu": await self.predict_rendement(
                parcelle_id,
                date_debut,
                etapes[-1]["date_fin"]
            )
        }

    async def _calculate_optimal_start_date(
        self,
        culture_type: CultureType,
        previsions_meteo: List[Dict[str, Any]]
    ) -> date:
        """Calcule la date optimale de début de cycle"""
        # TODO: Implémenter l'algorithme d'optimisation
        return date.today() + timedelta(days=7)

    async def _optimize_cycle_steps(
        self,
        culture_type: CultureType,
        date_debut: date,
        previsions_meteo: List[Dict[str, Any]]
    ) -> List[Dict[str, Any]]:
        """Optimise les étapes du cycle de culture"""
        # TODO: Implémenter l'algorithme d'optimisation
        return [
            {
                "etape": "Préparation",
                "date_debut": date_debut,
                "date_fin": date_debut + timedelta(days=7),
                "conditions_optimales": {
                    "temperature": "20-25°C",
                    "humidite": "60-70%"
                }
            },
            {
                "etape": "Croissance",
                "date_debut": date_debut + timedelta(days=7),
                "date_fin": date_debut + timedelta(days=37),
                "conditions_optimales": {
                    "temperature": "25-30°C",
                    "humidite": "70-80%"
                }
            },
            {
                "etape": "Récolte",
                "date_debut": date_debut + timedelta(days=37),
                "date_fin": date_debut + timedelta(days=44),
                "conditions_optimales": {
                    "temperature": "20-25°C",
                    "humidite": "50-60%"
                }
            }
        ]
